+++ projects/frontend/es-app/src/guis/menus/GuiMenuGamelistOptions.h
@@ -39,6 +39,8 @@ class GuiMenuGamelistOptions : public GuiMenuBase
       FlatFolders,
       MetaData,
       UpdateGamelist,
+      Delete,
+      DeleteScreeshots,
       MainMenu,
       Quit,
     };
@@ -58,6 +60,7 @@ class GuiMenuGamelistOptions : public GuiMenuBase
     std::shared_ptr<SwitchComponent> mFavoritesOnly;
     std::shared_ptr<SwitchComponent> mShowHidden;
     std::shared_ptr<TextComponent> mGame;
+    std::shared_ptr<TextComponent> mDeleteGame;
 
     //! Refresh first menu entry
     void RefreshGameMenuContext();